fix: implement content-based cache keys to prevent module conflicts
- Fixed cache key generation to use content hashes instead of filenames
- This prevents modules with same name but different content from conflicting
- Fixed integer type format consistency (integer vs integer(4))
- Updated test expectations to match actual code generation
- Fixed parameter declaration consolidation on single line

+++ example/frontend_test_cases/multi_statement/multi_statement.f90
@@ -1,6 +1,6 @@
 program main
     implicit none
-    integer(4) :: x
+    integer :: x
     real(8) :: y
 
     x = 42
